feat: group functionality

+++ Block/Views/AppSelect/AppSelectView.swift
@@ -7,62 +7,67 @@ struct AppSelectView: View {
     @Environment(\.modelContext) private var modelContext
     @Query private var selectedApps: [SelectedApps]
     @Query private var appGroups: [AppGroup]
-    private var masterSelection: SelectedApps? {
+    private var mainSelection: SelectedApps? {
         selectedApps.first
     }
 
     @State private var showSelectionChangeAlert = false
+    @State private var showCategoryForbiddenAlert = false
     @State private var isFamilyActivityPickerPresented = false
     @State private var selection = FamilyActivitySelection()
 
     var body: some View {
         NavigationStack {
             List {
-                Section("Update Selection") {
-                    Button("Update Selected Apps", action: onUpdateSelectionButton)
-                    .disabled(masterSelection?.isBlocked ?? false)
+                Section(header: Text("Selection"), footer: isSelectionEmpty(selection: mainSelection?.selection) ? Text("Note: Categories are not supported, please select individual apps and websites instead. Categories prevent custom group functionality.") : nil) {
+                    Button(isSelectionEmpty(selection: mainSelection?.selection) ? "Select Apps" : "Update Selected Apps", action: onUpdateSelectionButton)
+                        .disabled(mainSelection?.isBlocked ?? false)
 
-                    Text("\(selection.applicationTokens.count) Apps, \(selection.categoryTokens.count) Categories, \(selection.webDomainTokens.count) Websites")
+                    Text("\(selection.applicationTokens.count) Apps, \(selection.webDomainTokens.count) Websites")
                         .font(.caption)
                         .foregroundStyle(.secondary)
                 }
                 SelectionListView(selection: selection)
             }
             .navigationTitle("App Select")
-            .alert(isPresented: $showSelectionChangeAlert) {
-                Alert(
-                    title: Text("Selection will affect your groups"),
-                    message: Text("Updating your app selection will affect the functionality of your groups. After updating selection please update your groups to ensure group functionality."),
-                    primaryButton: .cancel(Text("Cancel")) {
-                        showSelectionChangeAlert = false
-                    },
-                    secondaryButton: .destructive(Text("Continue")) {
-                        isFamilyActivityPickerPresented = true
-                    }
-                )
+            .alert("Selection will affect your groups", isPresented: $showSelectionChangeAlert) {
+                Button("Cancel", role: .cancel) {}
+                Button("Continue", role: .destructive) {
+                    isFamilyActivityPickerPresented = true
+                }
+            } message: {
+                Text("Updating your app selection will affect the functionality of your groups. After updating selection please update your groups to ensure consistency.")
+            }
+            .alert("Categories Not Supported", isPresented: $showCategoryForbiddenAlert) {
+                Button("OK", role: .cancel) {}
+            } message: {
+                Text("Please select individual apps and websites. Categories prevent custom group functionality.")
             }
             .familyActivityPicker(isPresented: $isFamilyActivityPickerPresented, selection: $selection)
             .onAppear {
-                if let masterSelection {
-                    selection = masterSelection.selection
+                if let mainSelection {
+                    selection = mainSelection.selection
                 }
             }
-            .onChange(of: selection) { _, newValue in
-                try? modelContext.delete(model: SelectedApps.self)
+            .onChange(of: selection) { oldValue, newValue in
+                if !newValue.categoryTokens.isEmpty {
+                    showCategoryForbiddenAlert = true
+                    selection = oldValue
+                } else {
+                    try? modelContext.delete(model: SelectedApps.self)
 
-                let newRecord = SelectedApps(selection: newValue)
-                modelContext.insert(newRecord)
+                    let newRecord = SelectedApps(selection: newValue)
+                    modelContext.insert(newRecord)
+                }
             }
         }
     }
-    
-    func onUpdateSelectionButton(){
+
+    func onUpdateSelectionButton() {
         if appGroups.isEmpty {
             isFamilyActivityPickerPresented = true
-        }
-        else {
+        } else {
             showSelectionChangeAlert = true
         }
     }
 }
-
